pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "speakflow-local"
version = "0.1.0"
description = "Local-first macOS dictation app with Fn hold-to-talk and English/Hinglish modes."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "faster-whisper>=1.1.1",
  "numpy>=1.26.0",
  "pyobjc-framework-AVFoundation>=10.3",
  "pyobjc-framework-Cocoa>=10.3",
  "pyobjc-framework-Quartz>=10.3",
  "requests>=2.32.0",
  "sounddevice>=0.4.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0"
]
app = [
  "pyinstaller>=6.15.0"
]

[project.scripts]
speakflow = "whisper_flow.main:run"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_flow"]
